requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.13.0
supafunc==0.9.3
python-multipart==0.0.20
google-genai==1.3.0
google-auth==2.38.0